Reorganized the repos directory structure. Added CPC21 as OWL ontology. Extracted the transportation components of CPC and ICD-11
Aside from the directory structure re-organization, CPC21 was constructed as an OWL ontology using Protege's Cellfie and the available download of CPC21. cpc-transport-equipement was extracted using the robot utility. The The icd-transport-equipment.owl was extracted from the 2023-04-09 version of the whofic foundation using the robot utility. It includes 'Mode of transport of person injured in transport event', 'Counterpart in land transport crash', and the air, land, and water transport components of 'Objects, living things or substances involved in causing injury'.