HIRG Anatomy harmonization materials
The files are described in the document "HIRG Anatomy Harmonization.docx" located in the SharePoint folder "HIRG workspace/Working Groups/Content alignment - anatomy WG/"
whofic-2025-05-24.owl and whofic-2025-05-24.ttl are not included because of their size. They are not needed for the anatomy harmonization work, as the anatomy hierarchies extracted from them are sufficient.